requirements.txt
kivy==2.0.0
mysql-connector-python==8.0.33